Add Invoke-ComprehensiveWsusServerCleanup function
This function aims to perform all of the extra steps required to
maintain a pristine WSUS update catalog which Invoke-WsusServerCleanup
doesn't handle. It wraps around the latter, while performing numerous
optional tasks, to help trim your update approvals down to a minimum.

Note that this function was originally written as a personal script.
I've converted it to a function, but it's definitely below my usual
standard. Needs some love at some point in the future...