Install-Font: Update to use package memory streams with GlyphTypeface
Should fix #9.

This commit changes the usage of GlyphTypeface to read fonts from a
MemoryStream in a Package part instead of directly from the font file.

The underlying issue is the GlyphTypeface object doesn't release the
handle to the font file even after disposal and the GC runs, requiring
the PowerShell host to exit to release the handle. The issue is present
up to and including Windows 11 23H2, but is masked on newer releases
that default to using POSIX delete semantics.